+++ pycharm/src/main/java/rw/debugger/DropFrameActionHandler.java
@@ -2,12 +2,15 @@ package rw.debugger;
 
 import com.intellij.xdebugger.frame.XDropFrameHandler;
 import com.intellij.xdebugger.frame.XStackFrame;
-import com.jetbrains.python.debugger.PyStackFrame;
+import com.intellij.xdebugger.impl.XDebugSessionImpl;
+import com.jetbrains.python.debugger.*;
 import org.jetbrains.annotations.NotNull;
 import rw.session.Session;
 import rw.session.cmds.DropFrame;
 import rw.stack.Frame;
 import rw.stack.Stack;
+
+import java.util.Objects;
 
 public class DropFrameActionHandler implements XDropFrameHandler {
     private Session session;
@@ -22,11 +25,12 @@ public class DropFrameActionHandler implements XDropFrameHandler {
         PyStackFrame pyStackFrame = (PyStackFrame) frame;
 
         Frame reFrame = this.stack.getById(Long.valueOf(pyStackFrame.getFrameId()));
-        if (reFrame == null) {
+
+        if (reFrame == null || !reFrame.isReloadable()) {
             return false;
         }
 
-        return reFrame.getBack() != null;
+        return reFrame.getBack() != null && reFrame.getBack().isReloadable();
     }
 
     @Override