+++ pycharm/src/main/java/rw/session/events/DebuggerEvent.java
@@ -1,0 +1,8 @@
+package rw.session.events;
+
+import org.jetbrains.annotations.VisibleForTesting;
+import rw.handler.runConf.BaseRunConfHandler;
+
+abstract public class DebuggerEvent extends Event {
+    public static final String ID = null;
+}