+++ pycharm/src/main/java/rw/session/events/FrameDropped.java
@@ -9,13 +9,20 @@ import rw.preferences.PreferencesState;
 
 import java.awt.*;
 
-public class FrameDropped extends FrameEvent {
+public class FrameDropped extends FileEvent {
     public static final String ID = "FrameDropped";
+
+    @SerializedName("to_line")
+    private Integer toLine;
+
+    @SerializedName("from_line")
+    private Integer fromLine;
+
     private static final Logger LOGGER = Logger.getInstance(ModuleUpdate.class);
     @Override
     public void handle() {
         super.handle();
-        LOGGER.info("Handling FrameDopped " + String.format("(%s)", this.getFrameId()));
+        LOGGER.info("Handling FrameDopped");
 
         this.handler.getErrorHighlightManager().clearFile(this.getLocalPath());
 
@@ -23,8 +30,8 @@ public class FrameDropped extends FrameEvent {
 
         Color BLINK_COLOR = new Color(255, 0, 0, 60);
 
-        Blink blink = new Blink(this.handler.getProject(), this.getLocalPath(), this.getHandlerLineno(),
-                this.getLineno(),
+        Blink blink = new Blink(this.handler.getProject(), this.getLocalPath(), this.fromLine,
+                this.toLine,
                     BLINK_COLOR, -2, state.blinkDuration);
             Blinker.get().blink(blink);
     }